+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * Browser.setViewport
   * Assert.attribute and assert.not.attribute now match any element
   * Assert.href and assert.not.href now match any element
+  * Assert.visible message is now different when the element doesn't exist
 
 0.6.0 / 2018-04-13
 ==================